Reworked weather script (solid); minor formatting in rottentomatoes
Made weather script very solid by first having it try to precisely find
weather information by place name (input string). If that fails, the
data from geopy google geocoder supplies a lat,lng tuple which is used
as a backup. If all that somehow fails, an error is output to IRC with
a list of the closest matches.