Unobot AI, abstracted weather script into parts, added URL announce script
The biggest edits were working on the unobots AI and formatting of
cards. What it still needs is recognizing a played wild card (will
treat as a separate color and draw if played) and the playing of its
own special cards. The weather script was separated into parts for
allowing for two different parsing scripts: current conditions
(weather.py) and forecast.py. This means there is also a third script
(wunderground.py) for providing the API calls. Lastly, a simple URL
announce script was added to say the name of every properly formatted
URL linked to the irc bot (this inclues multiple URLs per irc event).